Identity contract: Separate tests into a different file
